% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOptimus.R
\name{parseH5ad}
\alias{parseH5ad}
\title{Parse an 5had file and extract the expression matrix and cell features (obs dataframe)}
\usage{
parseH5ad(
  h5ad_file,
  expression_matrix_path = "X",
  gene_id_path = "/var/ensembl_ids",
  cell_id_path = "/obs/CellID"
)
}
\arguments{
\item{h5ad_file}{Path to the h5ad file}

\item{expression_matrix_path}{Path to the raw counts expression matrix in the h5ad file}

\item{gene_id_path}{Path to the gene IDs in the h5ad file.}

\item{cell_id_path}{Path to the cell IDs in the h5ad file.}
}
\value{
A list containing the sparse matrix with genes in rows and cell barcodes in columns, and a dataframe of cell features if
the /obs path exists in the h5ad file.
}
\description{
This function tests if the incoming data is organized with genes in rows and cells in columns
or the transpose.  Output data is always genes in rows and cells in columns (Seurat/DropSeq format)
}
