% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOptimus.R
\name{parseOptimusH5ad}
\alias{parseOptimusH5ad}
\title{Parse an Optimus pipeline generated h5ad file.}
\usage{
parseOptimusH5ad(h5ad_file, min_transcripts = 20)
}
\arguments{
\item{h5ad_file}{Path to the h5ad file}

\item{min_transcripts}{Minimum number of transcripts required for a cell
barcode to be included in the output.}
}
\value{
A list(dge, cell_features) containing the sparse matrix with genes in
rows and cell barcodes in columns, and a dataframe of cell features needed
for nuclei selection.
}
\description{
This function parses the input, extracts the expression data as a sparse
matrix with cells in columns and gens in rows.  The cell level metadata
needed for nuclei selection is calculated / extracted from the obs dataframe.
}
\examples{
# Load the example h5ad file
h5ad_file <- system.file("extdata", "adata_example.h5ad.gz", package =
"DropSift")
r <- parseOptimusH5ad(h5ad_file, min_transcripts=0)
dge <- r$dge
cell_features <- r$cell_features
# Display output summary
dim(dge)  # Check dimensions of the parsed expression matrix
r$dge # The example DGE is 5 cells and 5 genes
#This cell feature set is processed to include the features needed for
DropSift.
head(cell_features)  # Preview the cell features dataframe.
}
