% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SvmNucleusCaller.R
\name{plotSvmNucleusCaller}
\alias{plotSvmNucleusCaller}
\title{S3 gorp that should not need to be documented but R CMD check complains if it is not.}
\usage{
plotSvmNucleusCaller(svmNucleusCaller)
}
\arguments{
\item{svmNucleusCaller}{an object of class SvmNucleusCaller}
}
\description{
S3 gorp that should not need to be documented but R CMD check complains if it is not.
}
