% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SvmNucleusCaller.R
\name{print.SvmNucleusCaller}
\alias{print.SvmNucleusCaller}
\title{Print the SvmNucleusCaller object summary}
\usage{
\method{print}{SvmNucleusCaller}(x, ...)
}
\arguments{
\item{x}{an object of class SvmNucleusCaller}

\item{...}{Other arguments that are ignored}
}
\description{
Print the SvmNucleusCaller object summary
}
