% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SvmNucleusCaller.R
\name{writeAnnotatedCellFeatures}
\alias{writeAnnotatedCellFeatures}
\alias{writeAnnotatedCellFeatures.SvmNucleusCaller}
\alias{writeAnnotatedCellFeatures.default}
\title{Write annotated cell features from an SvmNucleusCaller object}
\usage{
writeAnnotatedCellFeatures(svmNucleusCaller, outFile)

\method{writeAnnotatedCellFeatures}{SvmNucleusCaller}(svmNucleusCaller, outFile)

\method{writeAnnotatedCellFeatures}{default}(svmNucleusCaller, outFile)
}
\arguments{
\item{svmNucleusCaller}{An object of class \code{SvmNucleusCaller}.}

\item{outFile}{Character scalar giving the output path.}
}
\value{
Invisibly returns \code{outFile}.
}
\description{
Write the final annotated cell-feature table from an \code{SvmNucleusCaller}
object to a tab-separated file.  This contains the original inputs, plus
gene module scores, labels for each cell barcode if it is an exemplar class,
etc.
}
