% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SvmNucleusCaller.R
\name{getCBRBArgs}
\alias{getCBRBArgs}
\title{S3 gorp that should not need to be documented but R CMD check complains if it is not.}
\usage{
getCBRBArgs(svmNucleusCaller)
}
\arguments{
\item{svmNucleusCaller}{an object of class SvmNucleusCaller(useCBRBFeatures=FALSE)}
}
\description{
S3 gorp that should not need to be documented but R CMD check complains if it is not.
}
