% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SvmNucleusCaller.R
\name{plotSvmNucleusCaller}
\alias{plotSvmNucleusCaller}
\alias{plotSvmNucleusCaller.SvmNucleusCaller}
\title{Send a standard set of plots (3 pages) to the current graphics device}
\usage{
plotSvmNucleusCaller(svmNucleusCaller)

\method{plotSvmNucleusCaller}{SvmNucleusCaller}(svmNucleusCaller)
}
\arguments{
\item{svmNucleusCaller}{an object of class SvmNucleusCaller}
}
\value{
the input object, invisibly
}
\description{
Send a standard set of plots (3 pages) to the current graphics device
}
\examples{
# Load example dataset
data(svmNucleusCallerInputs)
set.seed(1)  # Set seed for reproducibility

# Create an SvmNucleusCaller object
svmNucleusCaller <- SvmNucleusCaller(
    cellFeatures = svmNucleusCallerInputs$cellFeatures,
    dgeMatrix = svmNucleusCallerInputs$dgeMatrix,
    datasetName = "example_dataset",
    useCBRBFeatures = FALSE,
    forceTwoClusterSolution = FALSE
)

# View summary
print(svmNucleusCaller)

# Plot the results, directing the outputs to a temp file.
outFile=tempfile(fileext=".pdf")
pdf(outFile)
plotSvmNucleusCaller(svmNucleusCaller)
dev.off()
}
