% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SvmNucleusCaller.R
\name{plotSvmNucleusCaller}
\alias{plotSvmNucleusCaller}
\alias{plotSvmNucleusCaller.SvmNucleusCaller}
\title{Send a standard set of plots (3 pages) to the current graphics device}
\usage{
plotSvmNucleusCaller(svmNucleusCaller)

\method{plotSvmNucleusCaller}{SvmNucleusCaller}(svmNucleusCaller)
}
\arguments{
\item{svmNucleusCaller}{an object of class SvmNucleusCaller}
}
\value{
the input object, invisibly

the input object
}
\description{
Send a standard set of plots (3 pages) to the current graphics device

Send a standard set of plots (3 pages) to the current graphics device
}
