% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SvmNucleusCaller-data.R
\docType{data}
\name{svmNucleusCallerInputs}
\alias{svmNucleusCallerInputs}
\title{svmNucleusCallerInputs}
\format{
A list with 2 elements:
cellFeatures: a data frame of cell features
dgeMatrix: a sparse matrix of digital gene expression
}
\usage{
data(svmNucleusCallerInputs)
}
\description{
cell features and digital gene expression matrix for testing the SVM nucleus caller.
Data from a snRNA-seq experiment
2023-08-25_v3_Bamboo_18d_10X_RNAseq_Optiprep8000_CaH_17k_rxn1_dev_aug07_2024
}
\examples{
data(svmNucleusCallerInputs)
svmNucleusCaller = DropSift::SvmNucleusCaller(svmNucleusCallerInputs$cellFeatures,
svmNucleusCallerInputs$dgeMatrix,
datasetName = "v3_Bamboo_18d_10X_RNAseq_Optiprep8000_CaH_17k_rxn1",
useCBRBFeatures = FALSE)

}
\keyword{datasets}
